# Displayed weather report in dialo when click on marker

+++ app/src/main/java/com/bala/openweathermap/api/response/Sys.java
@@ -6,69 +6,69 @@ import io.realm.RealmObject;
 
 public class Sys extends RealmObject {
 
-	@SerializedName("country")
-	private String country;
+    @SerializedName("country")
+    private String country;
 
-	@SerializedName("sunrise")
-	private int sunrise;
+    @SerializedName("sunrise")
+    private int sunrise;
 
-	@SerializedName("sunset")
-	private int sunset;
+    @SerializedName("sunset")
+    private int sunset;
 
-	@SerializedName("id")
-	private int id;
+    @SerializedName("id")
+    private long id;
 
-	@SerializedName("type")
-	private int type;
+    @SerializedName("type")
+    private int type;
 
-	@SerializedName("message")
-	private double message;
+    @SerializedName("message")
+    private double message;
 
-	public void setCountry(String country){
-		this.country = country;
-	}
+    public void setCountry(String country) {
+        this.country = country;
+    }
 
-	public String getCountry(){
-		return country;
-	}
+    public String getCountry() {
+        return country;
+    }
 
-	public void setSunrise(int sunrise){
-		this.sunrise = sunrise;
-	}
+    public void setSunrise(int sunrise) {
+        this.sunrise = sunrise;
+    }
 
-	public int getSunrise(){
-		return sunrise;
-	}
+    public int getSunrise() {
+        return sunrise;
+    }
 
-	public void setSunset(int sunset){
-		this.sunset = sunset;
-	}
+    public void setSunset(int sunset) {
+        this.sunset = sunset;
+    }
 
-	public int getSunset(){
-		return sunset;
-	}
+    public int getSunset() {
+        return sunset;
+    }
 
-	public void setId(int id){
-		this.id = id;
-	}
+    public void setId(long id) {
+        this.id = id;
+    }
 
-	public int getId(){
-		return id;
-	}
+    public long getId() {
+        return id;
+    }
 
-	public void setType(int type){
-		this.type = type;
-	}
+    public void setType(int type) {
+        this.type = type;
+    }
 
-	public int getType(){
-		return type;
-	}
+    public int getType() {
+        return type;
+    }
 
-	public void setMessage(double message){
-		this.message = message;
-	}
+    public void setMessage(double message) {
+        this.message = message;
+    }
 
-	public double getMessage(){
-		return message;
-	}
+    public double getMessage() {
+        return message;
+    }
 }
